requirements.txt
# Python dependencies for Lazada data processing
pandas==1.3.5
numpy>=1.20.0
mysql-connector-python>=8.0.27
python-dotenv>=0.19.0
requests>=2.26.0
SQLAlchemy==1.4.46
psycopg2-binary==2.9.5
scikit-learn==1.0.2 